23DEV053 complete middleware and cookie

+++ models/registermodel.go
@@ -10,17 +10,6 @@ type Demo struct {
 	Lname    string `json:"lname"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
-	Number   int64  `json:"number"`
+	Number   int    `json:"number"`
 	Profile  string `json:"profile"`
 }
-
-// type Show struct {
-// 	gorm.Model
-// 	Id       int    `gorm:"primary_key" json:"id"`
-// 	Fname    string `json:"fname"`
-// 	Lname    string `json:"lname"`
-// 	Email    string `json:"email"`
-// 	Password string `json:"password"`
-// 	Number   int64  `json:"number"`
-// 	Profile  string `json:"profile"`
-// }
